Use a single Style node internally
Edges, Nodes and Clusters can all have the "style" attribute, but the
allowed inputs are different for each. Use the same trick as before of
exposing type-safe constructors in the builder, that map to a single
Style node internally.

Choosing not the handle the "tapered", "striped", "wedged" styles for
now, as these need to be set along with other attributes to be useful.