Remove the default attribute scope
A default attribute (my own terminology) is one not scoped to an Edge,
Node or Graph. They can be defined like:

digraph {
  style="filled"
  ...
}

In this example, style is a default attribute. However, "filled" is a
style that is applicable to Edges, Nodes and Clusters. So which one does
it apply to here? In practice, the answer seems to be Clusters, but we
can avoid this ambiguity entirely by forcing all attributes to be
scoped, that is:

digraph {
  graph [style="filled"]
  ...
}

With Graphviz like this, it is clear the the "filled" style is being set
for Clusters.